Updated to 24292

+++ org/gjt/sp/jedit/pluginmgr/Roster.java
@@ -38,7 +38,7 @@ import static org.gjt.sp.jedit.io.FileVFS.recursiveDelete;
 //}}}
 
 /**
- * @author $Id: Roster.java 24045 2015-09-03 20:15:15Z kerik-sf $
+ * @author $Id: Roster.java 24071 2015-09-16 19:51:55Z kerik-sf $
  */
 class Roster
 {
@@ -64,6 +64,12 @@ class Roster
 		int size)
 	{
 		addOperation(new Install(installed,url,installDirectory,size));
+	} //}}}
+
+	//{{{ addLoad() method
+	void addLoad(String path)
+	{
+		toLoad.add(path);
 	} //}}}
 
 	//{{{ getOperation() method
@@ -203,27 +209,29 @@ class Roster
 
 			toLoad.remove(this.jar);
 
-			// move JAR first
 			File jarFile = new File(this.jar);
 			File srcFile = new File(this.jar.substring(0, this.jar.length() - 4));
 
-			Log.log(Log.NOTICE,this,"Deleting " + jarFile);
-
-			boolean ok = jarFile.delete();
-			if (ok) 
+			if(jarFile.exists())
 			{
-				EditBus.send(new PluginUpdate(jarFile, PluginUpdate.REMOVED, false));	
-			}
+				Log.log(Log.NOTICE,this,"Deleting " + jarFile);
 
-			if(srcFile.exists())
-			{
-				ok &= recursiveDelete(srcFile);
-			}
+				boolean ok = jarFile.delete();
+				if (ok)
+				{
+					EditBus.send(new PluginUpdate(jarFile, PluginUpdate.REMOVED, false));
+				}
 
-			if(!ok)
-			{
-				String[] args = {this.jar};
-				GUIUtilities.error(comp,"plugin-manager.remove-failed",args);
+				if(srcFile.exists())
+				{
+					ok &= recursiveDelete(srcFile);
+				}
+
+				if(!ok)
+				{
+					String[] args = {this.jar};
+					GUIUtilities.error(comp,"plugin-manager.remove-failed",args);
+				}
 			}
 		} //}}}
